+++ pkg/cmd/service/service.go
@@ -4,86 +4,212 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"log"
 	"time"
 
-	"github.com/kelseyhightower/envconfig"
+	grpc_auth "github.com/grpc-ecosystem/go-grpc-middleware/auth"
 	"github.com/openfga/openfga/pkg/encoder"
 	"github.com/openfga/openfga/pkg/logger"
 	"github.com/openfga/openfga/pkg/telemetry"
 	"github.com/openfga/openfga/server"
-	"github.com/openfga/openfga/server/authentication"
-	"github.com/openfga/openfga/server/authentication/oidc"
-	"github.com/openfga/openfga/server/authentication/presharedkey"
+	"github.com/openfga/openfga/server/authn"
+	"github.com/openfga/openfga/server/authn/oidc"
+	"github.com/openfga/openfga/server/authn/presharedkey"
 	"github.com/openfga/openfga/server/middleware"
 	"github.com/openfga/openfga/storage"
 	"github.com/openfga/openfga/storage/caching"
 	"github.com/openfga/openfga/storage/memory"
 	"github.com/openfga/openfga/storage/postgres"
+	"github.com/spf13/viper"
 	"google.golang.org/grpc"
 )
 
 var (
-	ErrInvalidGRPCTLSConfig = errors.New("'OPENFGA_GRPC_TLS_CERT_PATH' and 'OPENFGA_GRPC_TLS_KEY_PATH' env variables must be set")
-	ErrInvalidHTTPTLSConfig = errors.New("'OPENFGA_HTTP_GATEWAY_TLS_CERT_PATH' and 'OPENFGA_HTTP_GATEWAY_TLS_KEY_PATH' env variables must be set")
+	ErrInvalidGRPCTLSConfig = errors.New("'grpc.tls.cert' and 'grpc.tls.key' configs must be set")
+	ErrInvalidHTTPTLSConfig = errors.New("'http.tls.cert' and 'http.tls.key' configs must be set")
 )
 
-type Config struct {
-	// If you change any of these settings, please update the documentation at https://github.com/openfga/openfga.dev/blob/main/docs/content/intro/setup-openfga.mdx
-	DatastoreEngine               string `default:"memory" split_words:"true"`
-	DatastoreConnectionURI        string `split_words:"true"`
-	DatastoreMaxCacheSize         int    `default:"100000" split_words:"true"`
-	ServiceName                   string `default:"openfga" split_words:"true"`
-	HTTPPort                      int    `default:"8080" split_words:"true"`
-	RPCPort                       int    `default:"8081" split_words:"true"`
-	MaxTuplesPerWrite             int    `default:"100" split_words:"true"`
-	MaxTypesPerAuthorizationModel int    `default:"100" split_words:"true"`
-	// ChangelogHorizonOffset is an offset in minutes from the current time. Changes that occur after this offset will not be included in the response of ReadChanges.
-	ChangelogHorizonOffset int `default:"0" split_words:"true" `
-	// ResolveNodeLimit indicates how deeply nested an authorization model can be.
-	ResolveNodeLimit uint32 `default:"25" split_words:"true"`
-	// RequestTimeout is a limit on the time a request may take. If the value is 0, then there is no timeout.
-	RequestTimeout time.Duration `default:"0s" split_words:"true"`
+// DatabaseConfig defines OpenFGA server configurations for database specific settings.
+type DatabaseConfig struct {
 
-	GRPCTLSEnabled  bool   `default:"false" envconfig:"GRPC_TLS_ENABLED"`
-	GRPCTLSCertPath string `envconfig:"GRPC_TLS_CERT_PATH"`
-	GRPCTLSKeyPath  string `envconfig:"GRPC_TLS_KEY_PATH"`
+	// Engine is the database storage engine to use (e.g. 'memory', 'postgres')
+	Engine string
+	URI    string
 
-	HTTPTLSEnabled  bool   `default:"false" envconfig:"HTTP_TLS_ENABLED"`
-	HTTPTLSCertPath string `envconfig:"HTTP_TLS_CERT_PATH"`
-	HTTPTLSKeyPath  string `envconfig:"HTTP_TLS_KEY_PATH"`
+	// MaxCacheSize is the maximum number of cache keys that the storage cache can store before evicting
+	// old keys. The storage cache is used to cache query results for various static resources
+	// such as type definitions.
+	MaxCacheSize int
+}
 
-	// Authentication. Possible options: none,preshared,oidc
-	AuthMethod string `default:"none" split_words:"true"`
+// GRPCConfig defines OpenFGA server configurations for grpc server specific settings.
+type GRPCConfig struct {
+	Enabled bool
+	Addr    string
+	TLS     TLSConfig
+}
 
-	// Shared key authentication
-	AuthPresharedKeys []string `default:"" split_words:"true"`
+// HTTPConfig defines OpenFGA server configurations for HTTP server specific settings.
+type HTTPConfig struct {
+	Enabled bool
+	Addr    string
+	TLS     TLSConfig
 
-	// OIDC authentication
-	AuthOIDCIssuer   string `default:"" split_words:"true"`
-	AuthOIDCAudience string `default:"" split_words:"true"`
+	// UpstreamRequestTimeout is the timeout duration for proxying HTTP requests upstream
+	// to the grpc endpoint.
+	UpstreamRequestTimeout time.Duration
 
-	// Logging. Possible options: text,json
-	LogFormat string `default:"text" split_words:"true"`
-
-	// CORS configuration
 	CORSAllowedOrigins []string `default:"*" split_words:"true"`
 	CORSAllowedHeaders []string `default:"*" split_words:"true"`
 }
 
-func GetServiceConfig() Config {
-	var config Config
+// TLSConfig defines configuration specific to Transport Layer Security (TLS) settings.
+type TLSConfig struct {
+	Enabled  bool
+	CertPath string
+	KeyPath  string
+}
 
-	if err := envconfig.Process("OPENFGA", &config); err != nil {
-		log.Fatalf("failed to process server config: %v", err)
+// AuthnConfig defines OpenFGA server configurations for authentication specific settings.
+type AuthnConfig struct {
+
+	// Method is the authentication method that should be enforced (e.g. 'none', 'preshared', 'oidc')
+	Method                   string
+	*AuthnOIDCConfig         `mapstructure:"oidc"`
+	*AuthnPresharedKeyConfig `mapstructure:"preshared"`
+}
+
+// AuthnOIDCConfig defines configurations for the 'oidc' method of authentication.
+type AuthnOIDCConfig struct {
+	Issuer   string
+	Audience string
+}
+
+// AuthnPresharedKeyConfig defines configurations for the 'preshared' method of authentication.
+type AuthnPresharedKeyConfig struct {
+
+	// Keys define the preshared keys to verify authn tokens against.
+	Keys []string
+}
+
+// LogConfig defines OpenFGA server configurations for log specific settings. For production we
+// recommend using the 'json' log format.
+type LogConfig struct {
+
+	// Format is the log format to use in the log output (e.g. 'text' or 'json')
+	Format string
+}
+
+// PlaygroundConfig defines OpenFGA server configurations for the Playground specific settings.
+type PlaygroundConfig struct {
+	Enabled bool
+	Port    int
+}
+
+// OpenFGAConfig defines server configurations specific to the OpenFGA server itself.
+type OpenFGAConfig struct {
+
+	// MaxTuplesPerWrite defines the maximum number of tuples per Write endpoint.
+	MaxTuplesPerWrite int
+
+	// MaxTypesPerAuthorizationModel defines the maximum number of type definitions per authorization model for the WriteAuthorizationModel endpoint.
+	MaxTypesPerAuthorizationModel int
+
+	// ChangelogHorizonOffset is an offset in minutes from the current time. Changes that occur after this offset will not be included in the response of ReadChanges.
+	ChangelogHorizonOffset int
+
+	// ResolveNodeLimit indicates how deeply nested an authorization model can be.
+	ResolveNodeLimit uint32
+}
+
+type Config struct {
+	// If you change any of these settings, please update the documentation at https://github.com/openfga/openfga.dev/blob/main/docs/content/intro/setup-openfga.mdx
+
+	OpenFGA    OpenFGAConfig
+	Database   DatabaseConfig
+	GRPC       GRPCConfig
+	HTTP       HTTPConfig
+	Authn      AuthnConfig
+	Log        LogConfig
+	Playground PlaygroundConfig
+}
+
+// DefaultConfig returns the OpenFGA server default configurations.
+func DefaultConfig() *Config {
+	return &Config{
+		OpenFGA: OpenFGAConfig{
+			MaxTuplesPerWrite:             100,
+			MaxTypesPerAuthorizationModel: 100,
+			ChangelogHorizonOffset:        0,
+			ResolveNodeLimit:              25,
+		},
+		Database: DatabaseConfig{
+			Engine:       "memory",
+			MaxCacheSize: 100000,
+		},
+		GRPC: GRPCConfig{
+			Enabled: true,
+			Addr:    ":8081",
+			TLS:     TLSConfig{Enabled: false},
+		},
+		HTTP: HTTPConfig{
+			Enabled:                true,
+			Addr:                   ":8080",
+			TLS:                    TLSConfig{Enabled: false},
+			UpstreamRequestTimeout: 3 * time.Second,
+			CORSAllowedOrigins:     []string{"*"},
+			CORSAllowedHeaders:     []string{"*"},
+		},
+		Authn: AuthnConfig{
+			Method: "none",
+		},
+		Log: LogConfig{
+			Format: "text",
+		},
+		Playground: PlaygroundConfig{
+			Enabled: false,
+			Port:    3000,
+		},
 	}
-	return config
+}
+
+// GetServiceConfig returns the OpenFGA server configuration based on the values provided in the server's 'config.yaml' file.
+// The 'config.yaml' file is loaded from '/etc/openfga', '$HOME/.openfga', or the current working directory. If no configuration
+// file is present, the default values are returned.
+func GetServiceConfig() (*Config, error) {
+
+	config := DefaultConfig()
+
+	viper.SetConfigName("config")
+	viper.SetConfigType("yaml")
+
+	configPaths := []string{"/etc/openfga", "$HOME/.openfga", "."}
+	for _, path := range configPaths {
+		viper.AddConfigPath(path)
+	}
+	viper.AutomaticEnv()
+
+	err := viper.ReadInConfig()
+	if err != nil {
+		_, ok := err.(viper.ConfigFileNotFoundError)
+		if ok {
+			// if the server config is not found then return the defaults
+			return config, nil
+		}
+
+		return nil, fmt.Errorf("failed to load server config: %w", err)
+	}
+
+	if err := viper.Unmarshal(config); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal server config: %w", err)
+	}
+
+	return config, nil
 }
 
 type service struct {
 	server        *server.Server
 	datastore     storage.OpenFGADatastore
-	authenticator authentication.Authenticator
+	authenticator authn.Authenticator
 }
 
 func (s *service) Close(ctx context.Context) error {
@@ -96,105 +222,105 @@ func (s *service) Run(ctx context.Context) error {
 	return s.server.Run(ctx)
 }
 
-func BuildService(config Config, logger logger.Logger) (*service, error) {
+func BuildService(config *Config, logger logger.Logger) (*service, error) {
 	tracer := telemetry.NewNoopTracer()
 	meter := telemetry.NewNoopMeter()
 	tokenEncoder := encoder.NewBase64Encoder()
 
 	var datastore storage.OpenFGADatastore
 	var err error
-	switch config.DatastoreEngine {
+	switch config.Database.Engine {
 	case "memory":
-		datastore = memory.New(tracer, config.MaxTuplesPerWrite, config.MaxTypesPerAuthorizationModel)
+		datastore = memory.New(tracer, config.OpenFGA.MaxTuplesPerWrite, config.OpenFGA.MaxTypesPerAuthorizationModel)
 	case "postgres":
 		opts := []postgres.PostgresOption{
 			postgres.WithLogger(logger),
 			postgres.WithTracer(tracer),
 		}
 
-		datastore, err = postgres.NewPostgresDatastore(config.DatastoreConnectionURI, opts...)
+		datastore, err = postgres.NewPostgresDatastore(config.Database.URI, opts...)
 		if err != nil {
 			return nil, fmt.Errorf("failed to initialize postgres datastore: %v", err)
 		}
 	default:
-		return nil, fmt.Errorf("storage engine '%s' is unsupported", config.DatastoreEngine)
+		return nil, fmt.Errorf("storage engine '%s' is unsupported", config.Database.Engine)
 	}
 
-	logger.Info(fmt.Sprintf("using '%v' storage engine", config.DatastoreEngine))
+	logger.Info(fmt.Sprintf("using '%v' storage engine", config.Database.Engine))
 
 	var grpcTLSConfig *server.TLSConfig
-	if config.GRPCTLSEnabled {
-		if config.GRPCTLSCertPath == "" || config.GRPCTLSKeyPath == "" {
+	if config.GRPC.TLS.Enabled {
+		if config.GRPC.TLS.CertPath == "" || config.GRPC.TLS.KeyPath == "" {
 			return nil, ErrInvalidGRPCTLSConfig
 		}
 		grpcTLSConfig = &server.TLSConfig{
-			CertPath: config.GRPCTLSCertPath,
-			KeyPath:  config.GRPCTLSKeyPath,
+			CertPath: config.GRPC.TLS.CertPath,
+			KeyPath:  config.GRPC.TLS.KeyPath,
 		}
-		logger.Info("TLS is enabled, serving grpc connections using the provided certificate")
+		logger.Info("grpc TLS is enabled, serving connections using the provided certificate")
 	} else {
-		logger.Warn("TLS is disabled, serving grpc connections using insecure plaintext")
+		logger.Warn("grpc TLS is disabled, serving connections using insecure plaintext")
 	}
 
 	var httpTLSConfig *server.TLSConfig
-	if config.HTTPTLSEnabled {
-		if config.HTTPTLSCertPath == "" || config.HTTPTLSKeyPath == "" {
+	if config.HTTP.TLS.Enabled {
+		if config.HTTP.TLS.CertPath == "" || config.HTTP.TLS.KeyPath == "" {
 			return nil, ErrInvalidHTTPTLSConfig
 		}
 		httpTLSConfig = &server.TLSConfig{
-			CertPath: config.HTTPTLSCertPath,
-			KeyPath:  config.HTTPTLSKeyPath,
+			CertPath: config.HTTP.TLS.CertPath,
+			KeyPath:  config.HTTP.TLS.KeyPath,
 		}
-		logger.Info("HTTP TLS enabled, serving connections using the provided certificate")
+		logger.Info("HTTP TLS is enabled, serving HTTP connections using the provided certificate")
 	} else {
-		logger.Warn("HTTP TLS is disabled, falling back to insecure plaintext")
+		logger.Warn("HTTP TLS is disabled, serving connections using insecure plaintext")
 	}
 
-	var authenticator authentication.Authenticator
-	switch config.AuthMethod {
+	var authenticator authn.Authenticator
+	switch config.Authn.Method {
 	case "none":
-		logger.Warn("using 'none' authentication")
-		authenticator = authentication.NoopAuthenticator{}
+		logger.Warn("authentication is disabled")
+		authenticator = authn.NoopAuthenticator{}
 	case "preshared":
 		logger.Info("using 'preshared' authentication")
-		authenticator, err = presharedkey.NewPresharedKeyAuthenticator(config.AuthPresharedKeys)
+		authenticator, err = presharedkey.NewPresharedKeyAuthenticator(config.Authn.Keys)
 	case "oidc":
 		logger.Info("using 'oidc' authentication")
-		authenticator, err = oidc.NewRemoteOidcAuthenticator(config.AuthOIDCIssuer, config.AuthOIDCAudience)
+		authenticator, err = oidc.NewRemoteOidcAuthenticator(config.Authn.Issuer, config.Authn.Audience)
 	default:
-		return nil, fmt.Errorf("unsupported authenticator type: %v", config.AuthMethod)
+		return nil, fmt.Errorf("unsupported authentication method '%v'", config.Authn.Method)
 	}
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize authenticator: %v", err)
 	}
 
 	interceptors := []grpc.UnaryServerInterceptor{
-		middleware.NewAuthenticationInterceptor(authenticator),
+		grpc_auth.UnaryServerInterceptor(middleware.AuthFunc(authenticator)),
 	}
 
 	openFgaServer, err := server.New(&server.Dependencies{
-		Datastore:    caching.NewCachedOpenFGADatastore(datastore, config.DatastoreMaxCacheSize),
+		Datastore:    caching.NewCachedOpenFGADatastore(datastore, config.Database.MaxCacheSize),
 		Tracer:       tracer,
 		Logger:       logger,
 		Meter:        meter,
 		TokenEncoder: tokenEncoder,
 	}, &server.Config{
-		ServiceName: config.ServiceName,
 		GRPCServer: server.GRPCServerConfig{
-			Addr:      config.RPCPort,
+			Addr:      config.GRPC.Addr,
 			TLSConfig: grpcTLSConfig,
 		},
 		HTTPServer: server.HTTPServerConfig{
-			Addr:               config.HTTPPort,
-			TLSConfig:          httpTLSConfig,
-			CORSAllowedOrigins: config.CORSAllowedOrigins,
-			CORSAllowedHeaders: config.CORSAllowedHeaders,
+			Enabled:                config.HTTP.Enabled,
+			Addr:                   config.HTTP.Addr,
+			TLSConfig:              httpTLSConfig,
+			UpstreamRequestTimeout: config.HTTP.UpstreamRequestTimeout,
+			CORSAllowedOrigins:     config.HTTP.CORSAllowedOrigins,
+			CORSAllowedHeaders:     config.HTTP.CORSAllowedHeaders,
 		},
-		ResolveNodeLimit:       config.ResolveNodeLimit,
-		ChangelogHorizonOffset: config.ChangelogHorizonOffset,
+		ResolveNodeLimit:       config.OpenFGA.ResolveNodeLimit,
+		ChangelogHorizonOffset: config.OpenFGA.ChangelogHorizonOffset,
 		UnaryInterceptors:      interceptors,
 		MuxOptions:             nil,
-		RequestTimeout:         config.RequestTimeout,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize openfga server: %v", err)