+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2022-06-27
+### Added
+* OpenFGA Playground support (#68)
+* CORS policy configuration (#65)
+
 ## [0.1.2] - 2022-06-20
 ### Added
 * Request validation middleware
@@ -24,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Memory storage adapter implementation
 * Early support for preshared key or OIDC authentication methods
 
-[Unreleased]: https://github.com/openfga/openfga/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/openfga/openfga/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/openfga/openfga/releases/tag/v0.1.4
 [0.1.2]: https://github.com/openfga/openfga/releases/tag/v0.1.2
 [0.1.1]: https://github.com/openfga/openfga/releases/tag/v0.1.1
 [0.1.0]: https://github.com/openfga/openfga/releases/tag/v0.1.0